Move egraph to another file
